bot/requirements.txt
aiogram==3.20.0.post0
aiosqlite
python-dotenv
